correcciones

+++ BodaKaoz/Ceremonia/views.py
@@ -19,7 +19,12 @@ def ceremoniaDashboardView(request, user_id , boda_id , ceremonia_id):
 
 	user_id = request.user
 	enamorado = Enamorado.objects.get(User_id=user_id)
-	boda2 = Boda.objects.get(Enamorado1_id=enamorado.id)
+	indices_decoraciones = []
+
+	try:
+		boda2 = Boda.objects.get(Enamorado1_id=enamorado.id)
+	except:
+		boda2 = Boda.objects.get(Enamorado2_id=enamorado.id)
 	
 
 	if str(boda2.id) != str(boda_id):
@@ -66,17 +71,16 @@ def ceremoniaDashboardView(request, user_id , boda_id , ceremonia_id):
 		ministros = Ministro.objects.all()
 		lugares = Lugar.objects.filter(tipo='ceremonia')
 		musicas = Musica.objects.all()
+		boda = Boda.objects.get(id=boda_id)
 		decoraciones = DecoracionCeremonia.objects.all()
-		decoraciones_car = DecoracionCeremoniaCarrito.objects.all()
+		decoraciones_car = DecoracionCeremoniaCarrito.objects.filter(CeremoniaEvento_id=ceremonia_id)
 
+		indices_decoraciones.clear()
 
-		if len(decoraciones_car) > 0:
-			decora = []
-			for decoracion in decoraciones:
-				for deco in decoraciones_car:
-					if decoracion.nombre != deco.Decoracion.nombre:
-						decora.append(decoracion)
-			decoraciones = decora
+		if decoraciones_car.count() > 0:
+			indices_decoraciones.clear()
+			for deco in decoraciones_car:
+				indices_decoraciones.append(deco.Decoracion.id)
 
 
 
@@ -94,63 +98,107 @@ def ceremoniaDashboardView(request, user_id , boda_id , ceremonia_id):
 			'decoraciones':decoraciones,
 			'decoraciones_car':decoraciones_car,
 			'flag_decoracion':flag_decoracion,
+			'indices_decoraciones': indices_decoraciones
 
 		}
 		return HttpResponse(template.render(ctx,request))
 
 	if request.method == 'POST':
 		ceremonia = CeremoniaEvento.objects.get(Boda_id=boda_id)
+		boda = Boda.objects.get(id=boda_id)
 		btn_value = request.POST.get('btn_value')
+		decoraciones_car = DecoracionCeremoniaCarrito.objects.filter(CeremoniaEvento_id=ceremonia_id)
+		indices_decoraciones.clear()
+
+		if decoraciones_car.count() > 0:
+			indices_decoraciones.clear()
+			for deco in decoraciones_car:
+				indices_decoraciones.append(deco.Decoracion.id)
+
 
 		if btn_value == 'add_ministro':
-			price = request.POST.get('price')
-			id_ministro = request.POST.get('id_ministro')
-			ceremonia.Ministro_id = id_ministro
-			ceremonia.precio = ceremonia.precio + int(price)
-			flag_ministro = True
-			ceremonia.save()
+			if not 	flag_ministro:
+				price = request.POST.get('price')
+				id_ministro = request.POST.get('id_ministro')
+				ceremonia.Ministro_id = id_ministro
+				ceremonia.precio = ceremonia.precio + int(price)
+				flag_ministro = True
+				boda.precio = boda.precio + int(price)
+				boda.save()
+				ceremonia.save()
+			else:
+				pass
 
 		if btn_value == 'add_lugar':
-			price = request.POST.get('price')
-			id_ministro = request.POST.get('id_lugar')
-			ceremonia.Lugar_id = id_ministro
-			ceremonia.precio = ceremonia.precio + int(price)
-			flag_lugar = True
-			ceremonia.save()
+			if not flag_lugar:
+				price = request.POST.get('price')
+				id_ministro = request.POST.get('id_lugar')
+				ceremonia.Lugar_id = id_ministro
+				ceremonia.precio = ceremonia.precio + int(price)
+				flag_lugar = True
+				boda.precio = boda.precio + int(price)
+				boda.save()
+				ceremonia.save()
+			else:
+				pass
 
 		if btn_value == 'add_musica':
-			price = request.POST.get('price')
-			id_musica = request.POST.get('id_musica')
-			ceremonia.Musica_id = id_musica
-			ceremonia.precio = ceremonia.precio + int(price)
-			flag_musica = True
-			ceremonia.save()
+			if not flag_musica:
+				price = request.POST.get('price')
+				id_musica = request.POST.get('id_musica')
+				ceremonia.Musica_id = id_musica
+				ceremonia.precio = ceremonia.precio + int(price)
+				flag_musica = True
+				boda.precio = boda.precio + int(price)
+				boda.save()
+				ceremonia.save()
+			else:
+				pass
 
 		if btn_value == 'add_fotos':
-			ceremonia.Fotos = True
-			flag_fotos = True
-			ceremonia.save()
+			if not flag_fotos:
+				ceremonia.Fotos = True
+				flag_fotos = True
+				ceremonia.save()
+			else:
+				pass
 
 		if btn_value == 'delete_ministro':
-			price = request.POST.get('price')
-			ceremonia.Ministro_id = None
-			ceremonia.precio = ceremonia.precio - int(price)
-			flag_ministro = False
-			ceremonia.save()
+			if flag_ministro:
+				price = request.POST.get('price')
+				ceremonia.Ministro_id = None
+				ceremonia.precio = ceremonia.precio - int(price)
+				flag_ministro = False
+				boda.precio = boda.precio - int(price)
+				boda.save()
+				ceremonia.save()
+			else:
+				pass
 
 		if btn_value == 'delete_lugar':
-			price = request.POST.get('price')
-			ceremonia.Lugar_id = None
-			ceremonia.precio = ceremonia.precio - int(price)
-			flag_lugar = False
-			ceremonia.save()
+			if flag_lugar:
+				price = request.POST.get('price')
+				ceremonia.Lugar_id = None
+				ceremonia.precio = ceremonia.precio - int(price)
+				flag_lugar = False
+				boda.precio = boda.precio - int(price)
+				boda.save()
+				ceremonia.save()
+			else:
+				pass
 
 		if btn_value == 'delete_musica':
-			price = request.POST.get('price')
-			ceremonia.Musica_id = None
-			ceremonia.precio = ceremonia.precio - int(price)
-			flag_musica = False
-			ceremonia.save()
+			if flag_musica:
+				price = request.POST.get('price')
+				ceremonia.Musica_id = None
+				ceremonia.precio = ceremonia.precio - int(price)
+				flag_musica = False
+				boda.precio = boda.precio - int(price)
+				boda.save()
+				ceremonia.save()
+			else:
+				pass
+
 
 		if btn_value == "add_decoracion":
 			cantidad = request.POST.get('cantidad')
@@ -158,22 +206,43 @@ def ceremoniaDashboardView(request, user_id , boda_id , ceremonia_id):
 			id_decoracion = request.POST.get('id_decoracion')
 			price = request.POST.get('price')
 			decoracion = DecoracionCeremonia.objects.filter(id__exact=id_decoracion)
-			decoracion_carrito = DecoracionCeremoniaCarrito(Decoracion=decoracion[0] , CeremoniaEvento = ceremonia , cantidad = cantidad)
-			ceremonia.precio = ceremonia.precio + (int(price) * int (cantidad))
-			ceremonia.save()
-			decoracion_carrito.save()
+
+			if int(id_decoracion) not in indices_decoraciones:
+				decoracion_carrito = DecoracionCeremoniaCarrito(Decoracion=decoracion[0] , CeremoniaEvento = ceremonia , cantidad = cantidad , subtotal=(int(price) * int (cantidad)))
+				ceremonia.precio = ceremonia.precio + (int(price) * int (cantidad))
+				boda.precio = boda.precio + (int(price) * int(cantidad))
+				boda.save()
+				ceremonia.save()
+				decoracion_carrito.save()
+				indices_decoraciones.clear()
+
+				if decoraciones_car.count() > 0:
+					for deco in decoraciones_car:
+						indices_decoraciones.append(deco.Decoracion.id)
+
+			else:
+				pass
 
 		if btn_value == "delete_decoracion":
-			comida_id = request.POST.get('id_decoracion')
-			decoracion_carrito_id = request.POST.get('id_decoracion_carrito')
-			decoracioncarrito = DecoracionCeremoniaCarrito.objects.filter(id__exact=decoracion_carrito_id)
-			cantidad = decoracioncarrito[0].cantidad
-			decoracioncarrito.delete()
-			price = request.POST.get('price')
-			print(cantidad)
-			print(price)
-			ceremonia.precio = ceremonia.precio - (int(cantidad)  * int(price))
-			ceremonia.save()
+			decoracion_id = request.POST.get('id_decoracion')
+
+			if int(decoracion_id) in indices_decoraciones:
+				decoracion_carrito_id = request.POST.get('id_decoracion_carrito')
+				decoracioncarrito = DecoracionCeremoniaCarrito.objects.filter(id__exact=decoracion_carrito_id)
+				cantidad = decoracioncarrito[0].cantidad
+				decoracioncarrito.delete()
+				price = request.POST.get('price')
+				ceremonia.precio = ceremonia.precio - (int(cantidad)  * int(price))
+				boda.precio = boda.precio - (int(price) * int(cantidad))
+				boda.save()
+				ceremonia.save()
+				indices_decoraciones.clear()
+
+				if decoraciones_car.count() > 0:
+					for deco in decoraciones_car:
+						indices_decoraciones.append(deco.Decoracion.id)
+			else:
+				pass
 
 		if btn_value == 'delete_fotos':
 			ceremonia.Fotos = False
@@ -185,15 +254,13 @@ def ceremoniaDashboardView(request, user_id , boda_id , ceremonia_id):
 		musicas = Musica.objects.all()
 		lugares = Lugar.objects.filter(tipo='ceremonia')
 		decoraciones = DecoracionCeremonia.objects.all()
-		decoraciones_car = DecoracionCeremoniaCarrito.objects.all()
+		decoraciones_car = DecoracionCeremoniaCarrito.objects.filter(CeremoniaEvento_id=ceremonia_id)
 
-		if len(decoraciones_car) > 0:
-			decora = []
-			for decoracion in decoraciones:
-				for deco in decoraciones_car:
-					if decoracion.nombre != deco.Decoracion.nombre:
-						decora.append(decoracion)
-			decoraciones = decora
+		indices_decoraciones.clear()
+
+		if decoraciones_car.count() > 0:
+			for deco in decoraciones_car:
+				indices_decoraciones.append(deco.Decoracion.id)
 
 		template = get_template('Ceremonia/ceremonia.html')
 		ctx={
@@ -209,6 +276,7 @@ def ceremoniaDashboardView(request, user_id , boda_id , ceremonia_id):
 			'musicas' : musicas,
 			'decoraciones':decoraciones,
 			'decoraciones_car':decoraciones_car,
+			'indices_decoraciones': indices_decoraciones
 
 
 		}
